break out call, note, and search functions into their own concerns
