+++ app/models/concerns/urgency.rb
@@ -25,4 +25,11 @@ module Urgency
       end
     end
   end
+
+  private
+
+  def recent_history_tracks
+    history_tracks.select { |ht| ht.updated_at > 6.days.ago }
+  end
+
 end
